+++ inc/macros.h
@@ -29,7 +29,6 @@
 	# define		PROG_NAME	"EstDis"
 	# define		PROG_YEAR	"2015"
 	# define		PROG_VERS	"2.0"
-	# define		OPT_LIST	"a:b:c:f:hi:j:k:lo:p:r:s:V:vx"
 
 	# define		ESTDIS_PATH	"./"
 	# define		LICENSE_RPATH	"../COPYING.txt"
