+++ modules/Makefile
@@ -11,6 +11,7 @@ MODULES_DIR = $(ESTDIS_DIR)/modules/
 ABOUT_DIR = $(MODULES_DIR)/about/
 CALC_DIR = $(MODULES_DIR)/calc/
 CTRL_DIR = $(MODULES_DIR)/ctrl/
+DATA_DIR = $(MODULES_DIR)/data/
 PARSE_DIR = $(MODULES_DIR)/parse/
 TUI_DIR = $(MODULES_DIR)/tui/
 
@@ -24,6 +25,7 @@ all:
 	cd $(ABOUT_DIR) && $(MAKE) && cd ..
 	cd $(CALC_DIR) && $(MAKE) && cd ..
 	cd $(CTRL_DIR) && $(MAKE) && cd ..
+	cd $(DATA_DIR) && $(MAKE) && cd ..
 	cd $(PARSE_DIR) && $(MAKE) && cd ..
 	cd $(TUI_DIR) && $(MAKE) && cd ..
 
@@ -32,5 +34,6 @@ clean:
 	cd $(ABOUT_DIR) && $(MAKE) clean && cd ..
 	cd $(CALC_DIR) && $(MAKE) clean && cd ..
 	cd $(CTRL_DIR) && $(MAKE) clean && cd ..
+	cd $(DATA_DIR) && $(MAKE) clean && cd ..
 	cd $(PARSE_DIR) && $(MAKE) clean && cd ..
 	cd $(TUI_DIR) && $(MAKE) clean && cd ..
