+++ modules/data/Makefile
@@ -1,0 +1,21 @@
+# -*- MakeFile -*-
+
+# MACRO = substitute with this
+
+# directories
+
+ESTDIS_DIR = ../../
+
+MODULES_DIR = $(ESTDIS_DIR)/modules/
+DATA_DIR = $(MODULES_DIR)/data/
+OBJ_DIR = $(DATA_DIR)/obj/
+TST_DIR = $(DATA_DIR)/tst/
+
+# target: dependencies
+#	action
+
+all:
+	cd $(OBJ_DIR) && $(MAKE) && cd ..
+
+clean:
+	cd $(OBJ_DIR) && $(MAKE) clean && cd ..
