+++ modules/parse/src/parse.c
@@ -15,22 +15,85 @@
  * to the Free Software Foundation, Inc.				      *
  ******************************************************************************/
 
+
+/******************************************************************************
+ ******| include |*************************************************************
+ ******************************************************************************/
 	#include <getopt.h>
-		/* need: int64_t & INT64_MIN & INT64_MAX  & SCNi64 */
+		/* int64_t & INT64_MIN & INT64_MAX & SCNi64 */
 	#include <inttypes.h>
-		/* need: INFINITY */
+		/* INFINITY */
 	#include <math.h>
+		/* printf() */
 	#include <stdio.h>
+		/* exit? */
 	#include <stdlib.h>
-	#include <string.h>
 
+		/* print functions */
 	#include "about.h"
-	#include "parse.h"
-
-	#include "global_ext.h"
-	#include "macros.h"
+		/* global variables */
+	#include "data.h"
 
 
+/******************************************************************************
+ ******| options |*************************************************************
+ ******************************************************************************/
+	enum	Parse_Short_Option {
+		PARSE_S_HELP =		'h',
+		PARSE_S_LICENSE =	'l',
+		PARSE_S_START =		's',
+		PARSE_S_VERBOSE =	'V',
+		PARSE_S_VERSION =	'v',
+		PARSE_S_EXIT =		'x',
+
+		PARSE_S_CPRIGHT =	10000,	/* No short option */
+		PARSE_S_WARRANTY,		/* No short option */
+
+		PARSE_S_PARAM_c,		/* No short option */
+		PARSE_S_PARAM_n,		/* No short option */
+		PARSE_S_PARAM_N,		/* No short option */
+		PARSE_S_PARAM_r,		/* No short option */
+
+		PARSE_S_PARAM_a,		/* No short option */
+		PARSE_S_PARAM_b,		/* No short option */
+		PARSE_S_PARAM_p,		/* No short option */
+		PARSE_S_PARAM_x0,		/* No short option */
+		PARSE_S_PARAM_x1,		/* No short option */
+		PARSE_S_PARAM_x2,		/* No short option */
+		PARSE_S_PARAM_beta,		/* No short option */
+		PARSE_S_PARAM_lambda		/* No short option */
+	};
+
+	# define	OPT_LIST		"hls:V:vx"
+
+	# define	PARSE_L_HELP 		"help"
+	# define	PARSE_L_LICENSE 	"license"
+	# define	PARSE_L_START 		"start"
+	# define	PARSE_L_VERBOSE 	"Verbose"
+	# define	PARSE_L_VERSION 	"version"
+	# define	PARSE_L_EXIT 		"exit"
+
+	# define	PARSE_L_CPRIGHT 	"copyright"
+	# define	PARSE_L_WARRANTY 	"warranty"
+
+	# define	PARSE_L_PARAM_c 	"param-c"
+	# define	PARSE_L_PARAM_n 	"param-n"
+	# define	PARSE_L_PARAM_N 	"param-N"
+	# define	PARSE_L_PARAM_r 	"param-r"
+
+	# define	PARSE_L_PARAM_a 	"param-a"
+	# define	PARSE_L_PARAM_b 	"param-b"
+	# define	PARSE_L_PARAM_p 	"param-p"
+	# define	PARSE_L_PARAM_x0 	"param-x0"
+	# define	PARSE_L_PARAM_x1 	"param-x1"
+	# define	PARSE_L_PARAM_x2 	"param-x2"
+	# define	PARSE_L_PARAM_beta 	"param-beta"
+	# define	PARSE_L_PARAM_lambda 	"param-lambda"
+
+
+/******************************************************************************
+ ******| static functions |****************************************************
+ ******************************************************************************/
 static	int64_t	parse_int	(const char *optstr, const char *optarg,
 				int64_t *param, int64_t m, int64_t M);
 
@@ -38,6 +101,9 @@ static	int64_t	parse_dbl	(const char *optstr, const char *optarg,
 				double *param, double m, double M);
 
 
+/******************************************************************************
+ ******| parser |**************************************************************
+ ******************************************************************************/
 void	parser	(int argc, char *argv[])
 {
 	int	opt =		0;
@@ -90,7 +156,7 @@ void	parser	(int argc, char *argv[])
 
 		case PARSE_S_VERBOSE:
 			if (parse_int(PARSE_L_VERBOSE, optarg, &flag_V,
-						VERBOSE_Q, VERBOSE_MAX)) {
+						VERBOSE_0, VERBOSE_MAX)) {
 				exit(EXIT_FAILURE);
 			}
 			break;
@@ -210,6 +276,9 @@ void	parser	(int argc, char *argv[])
 }
 
 
+/******************************************************************************
+ ******| static functions |****************************************************
+ ******************************************************************************/
 static	int64_t	parse_int	(const char *optstr, const char *optarg,
 				int64_t *param, int64_t m, int64_t M)
 {
